requirements.txt
aiofiles==24.1.0
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.3
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email-validator==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.18.1
filelock==3.13.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
functions==0.7.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
python-jose==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
middleware==1.2.3
mongoengine==0.28.2
motor==3.4.0
numpy==1.26.1
openpyxl==3.1.2
orjson==3.10.3
packaging==24.1
passlib==1.7.4
pip==24.1
platformdirs==4.2.0
pluggy==1.5.0
pyasn1==0.6.0
pydantic==2.4.2
pydantic-core==2.10.1
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.3.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytups==0.86.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
schemas==0.7.1
setuptools==63.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
tqdm==4.66.1
typer==0.12.3
typing-extensions==4.11.0
ujson==5.10.0
urllib3==2.0.7
uvicorn==0.29.0
virtualenv==20.25.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
XlsxWriter==3.1.9
